Set up a cart in the API context with an in-memory store
Normally carts are stored in a "session" backed by a cookie. We want to
be able to use carts to do things like calculate shipping, but putting
it in a Cookie doesn't make sense in the REST API context.

The WC_Session class turns out to be a basic in-memory key/value store
that works well for this use case.

We're loading the cart in the API Server init so that it's available to
all endpoints. We may want to consider doing this on a case by case
basis though.